requirements.txt
rich>=13.7.0
textual>=0.52.1
requests>=2.31.0
psutil>=5.9.8
pynvml>=11.5.0
pyyaml>=6.0.1
argparse>=1.4.0
pytz>=2024.1